try to come up with good numerical examples
the main function generate a figure of flow inside a chip confined by smoothen poly, with regular rectangle island, and a tank like island inside. not very stable in tems of changing number of panels (corner vs. straight side). This may have something to do with choice of layer potential, or it could be an issue with density solve.  densities on confined geometry and rectangle are smooth, but density on confined geometry seems to be very high...(1e+07). density on tank island is more giggly wiggly like